Revert "Merge branch 'SKR-PRO_v1_1' into PR_SKR_PRO_v11_config"
This reverts commit cd02722fb7fb8eb8c20c4de96308d004cf3fcdd7, reversing
changes made to b107fb758aeffddeedc1554d20c294e153966c8a.

+++ Src/diskio.c
@@ -408,7 +408,7 @@ DRESULT disk_read (
     BYTE drv,            /* Physical drive nmuber (0) */
     BYTE *buff,            /* Pointer to the data buffer to store read data */
     DWORD sector,        /* Start sector number (LBA) */
-    UINT count            /* Sector count (1..255) */
+    BYTE count            /* Sector count (1..255) */
 )
 {
     if (drv || !count) return RES_PARERR;
